Fixes after code review

+++ CVapp.Domain/Models/Content/Money.cs
@@ -7,7 +7,7 @@
         public string TransactionAccount { get; set; }
         public string Category { get; set; }
         public string Item { get; set; }
-        public float Sum { get; set; }
+        public decimal Sum { get; set; }
         public string TransactionType { get; set; }
         public string Currency { get; set; }
     }
